Refactor into separate methods.
The SQL based driver can cheat here as it does not need to heed the
permissions. While it pretends to be "john" it edits the permissions
of system shares to which the real "john" would not be granted any
access.

For the Kolab IMAP backend there is no such shortcut since it will
always enforce the permissions. So for testing it must be possible to
switch the users.