Compare single permission settings rather than complete objects.
The Kolab permission backend is very different from the way this is
handled with SQL as a backend. The old testing did compare complete
permission objects but it is hard to get this to align for both
backends. Unset permissions are not present for example and the
permission object itself is not of type Horde_Perms_Permission but
Horde_Perms_Permission_Kolab.

Instead of comparing complete permission objects the tests now verify
the expected permission settings one by one.